+++ src/Command/DumpCommand.php
@@ -13,11 +13,13 @@ namespace Nelmio\ApiDocBundle\Command;
 
 use Nelmio\ApiDocBundle\Render\Html\AssetsMode;
 use Nelmio\ApiDocBundle\Render\RenderOpenApi;
+use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
+#[AsCommand(name: 'nelmio:apidoc:dump')]
 final class DumpCommand extends Command
 {
     private RenderOpenApi $renderOpenApi;
@@ -47,6 +49,18 @@ final class DumpCommand extends Command
         $availableFormats = $this->renderOpenApi->getAvailableFormats();
         $this
             ->setDescription('Dumps documentation in OpenAPI format to: '.implode(', ', $availableFormats))
+            ->setHelp(
+                <<<'EOF'
+                    The <info>%command.name%</info> command dumps the OpenAPI documentation in the specified format to the standard output.
+                    You can specify the "area" option to select a specific documentation area (if multiple areas are defined).
+
+                      <info>php %command.full_name%</info>  
+                      <info>php %command.full_name% --area MyAreaName</info>
+
+                    <href=https://symfony.com/bundles/NelmioApiDocBundle/current/commands.html>Read the documentation here.</href>
+
+                    EOF
+            )
             ->addOption('area', '', InputOption::VALUE_OPTIONAL, '', 'default')
             ->addOption(
                 'format',